+++ pybads/testing/bads/poll/test_poll_mads.py
@@ -8,7 +8,13 @@ def poll_mads_2n_ones_test():
     search_mesh_size = 9.7656e-4
     mesh_size = 1.0
     D = 3
-    B = poll_mads_2n(D, poll_scale, search_mesh_size, mesh_size)
+    B = poll_mads_2n(
+        D,
+        poll_scale,
+        search_mesh_size,
+        mesh_size,
+        rng=np.random.default_rng(0),
+    )
     assert B.shape[0] == 6 and B.shape[1] == D
 
 
@@ -17,5 +23,11 @@ def poll_mads_2n_test():
     search_mesh_size = 9.7656e-4
     mesh_size = 0.0312
     D = 3
-    B = poll_mads_2n(D, poll_scale, search_mesh_size, mesh_size)
+    B = poll_mads_2n(
+        D,
+        poll_scale,
+        search_mesh_size,
+        mesh_size,
+        rng=np.random.default_rng(0),
+    )
     assert B.shape[0] == 6 and B.shape[1] == D