+++ pybads/testing/bads/search/test_search.py
@@ -72,7 +72,7 @@ def test_search():
     es_iter = bads.options["n_search_iter"]
     mu = int(bads.options["n_search"] / es_iter)
     lamb = mu
-    search_es = ESSearchWM(mu, lamb, bads.options)
+    search_es = ESSearchWM(mu, lamb, bads.options, rng=bads.rng)
     us, z = search_es(
         bads.u, lb, ub, bads.function_logger, gp, bads.optim_state, True, None
     )
@@ -80,7 +80,7 @@ def test_search():
     assert us.size == 3 and (np.isscalar(z) or z.size == 1)
     assert np.all(gp.y >= z)
 
-    search_es = ESSearchELL(mu, lamb, bads.options)
+    search_es = ESSearchELL(mu, lamb, bads.options, rng=bads.rng)
     us, z = search_es(
         bads.u, lb, ub, bads.function_logger, gp, bads.optim_state, True, None
     )
@@ -96,7 +96,7 @@ def test_search_selection_mask():
         D,
         get_pybads_option_dir_path(),
     )
-    search_es = ESSearchWM(mu, lamb, options)
+    search_es = ESSearchWM(mu, lamb, options, rng=np.random.default_rng(0))
     mask = search_es._get_selection_idx_mask_(mu, lamb)
     assert np.sum(mask) == 885072
     assert np.min(mask + 1) == 1
@@ -115,7 +115,9 @@ def test_search_hedge():
     bads.options["fun_eval_start"] = 10
     gp, Ns_gp, sn2hpd, hyp_dict = bads._init_optimization_()
 
-    search_hedge = ESSearchHedge(bads.options["search_method"], bads.options)
+    search_hedge = ESSearchHedge(
+        bads.options["search_method"], bads.options, rng=bads.rng
+    )
 
     us, z = search_hedge(
         bads.u, lb, ub, bads.function_logger, gp, bads.optim_state